implemented and tested new waiting validators set feature
