+++ charts/polkadot-watcher-csv-exporter/Chart.yaml
@@ -1,4 +1,4 @@
 description: Polkadot Watcher
 name: polkadot-watcher-csv-exporter
-version: v0.6.0
+version: v0.7.0
 apiVersion: v2
