+++ charts/polkadot-watcher-csv-exporter/values.yaml
@@ -2,7 +2,7 @@ environment: production
 
 image:
   repo: web3f/polkadot-watcher-csv-exporter
-  tag: v0.6.0
+  tag: v0.7.0
 
 config:
   endpoint: "wss://kusama-rpc.polkadot.io"
