+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "polkadot-watcher-csv-exporter",
-  "version": "0.6.0",
+  "version": "0.7.0",
   "description": "Monitor events on Polkadot networks and export a csv dataset",
   "repository": "git@github.com:w3f/polkadot-watcher-csv-exporter.git",
   "author": "W3F Infrastructure Team <devops@web3.foundation>",
